add ‘dmenushader’ script and ‘Grayscale’ shader
‘dmenushader’ allows the user to select GLSL shaders for picom from
dmenu. Shaders are stored in ‘$XDG_DATA_HOME/shaders’ as *.glsl files
and may be paired with a *.sh script to be executed on shader selection.

Includes a ‘Grayscale.glsl’ shader which applies a grayscale filter to
all windows, as well as a ‘Grayscale.sh’ script which applies a
grayscale filter to the background image using ImageMagick and resets it
on termination.