[XCompose] add Christian crosses.
Having several such crosses readiy available makes it trivial to refer
to Christianity without having to resort to controversial contractions
like ‘χianity’. The various crosses also allow one to make quick
reference to Orthodox Christians (by using the Suppedaneum cross, “☦”),
to specific orders (such as the Knights Hospitaller, by using the
Maltese Cross) et cetera.

One of the problems with this is that most fonts don’t display such
crosses accurately. As such, the convention (as declared by me) is to
use these code points as they are named by the Unicode Consortium. For
example, the Maltese Cross (✠) is often displayed as a croix pattée,
but is still treated as a Maltese Cross. Less commonly, the Croix de
Lorraine (☨) is displayed as a Patriarchal cross.

The Globus Cruciger is also present in the ‘Planet Symbols’ section,
but, given the symbol's origin in mediaeval Christianity (wherein it
represents Christ's dominion over the orb of the world), it makes sense
to have it here as well, under a different key combination.