[XCompose] add Christograms
As we've added Christian crosses, it makes sense to add the Chrismon (☧)
and the Staurogram (⳨), even though the latter is not too common.

Like the previous commit, this one makes it easier to refer to a
specific denomination, as the Chrismon topped the Labara (vexilla) of
Constantine's army, which associates it with the early Catholic Church.

Nevertheless, it would be nice to have more Christograms, such as the
ΙΧ, ΙΗ, ICXC, or IHS/JHS monograms, but I can't find any in Unicode.
Other desireable additions would be the Good Shepherd, the Anchored
Cross, the Eye of Providence, the Lily Crucifix et cetera. The Dove and
the Shamrock already exist.

(The Ichthys would be cool as well… Should I contact the Unicode
Consortium about this?)