[XCompose] add more Crux immissa variants
As it turns out, the previous Crux immissa was an ***emoji***. As this
is absolutely unacceptable, I've added more crosses and made the emoji
one more difficult to type by accident.

And yes, I know that the Suppedaneum cross is an emoji as well, but
there's no way to get around that one. And no, I will not add all the
Greek crosses from the ‘Geometric Shapes Extended’ block. I haven't the
time.