[genpass] allow custom size and encoding
By default, genpass would only generate 24-byte passphrases, and only
encode them in base64. This commit preserves these as the defaults, and
adds the option to specify — as the first argument — the number of bytes
(24 if it's an empty string), and — as the subsequent arguments — the
formatting command (“base64 -w0” if it's an empty string).

In addition, a new keybinding was added to dwm to allow for the
generation of base58-encoded 24-byte passhrases in order to make strings
that are human-readable and url-friendly. (aur/base-ex-cli must be
installed.)