[.gitconfig] set checkout.defaultRemote to origin
By default, when using `git checkout <branch>` for a missing <branch>
which is found in exactly one remote, it will create it and
automatically track <remote>/<branch>. However, if <branch> exists in
more than one remote, this commit causes it to default to `origin`.

See git-checkout(1), git-config(1), and git-switch(1) for more info.