dwmbar: display number of available package updates
We use the “checkupdates” script from the “pacman-contrib” package to
safely check for available package updates. The execution of this script
is managed by a systemd timer (see `man 5 systemd.timer`) and a
corresponding service. The default “checkupdates” database directory is
changed to ${XDG_CACHE_HOME} to persist across reboots. The timer — and
not the service — should be enabled for this to work. To do that, run:

```sh
systemctl --user daemon-reload
systemctl --user enable --now checkupdates.timer
```

The “dwmbar” script is changed to include a “updates()” function, which
will only display anything if one or more packages can be upgraded. In
that case, the number of packages will be indicated. In addition, if any
kernel updates are available, a penguin icon will be displayed next to
the number of packages.

Finally, the README has been updated to tell the user to enable any
systemd timers.