nvim: use NvChad as a base configuration
NvChad[1] is a nice and fast base configuration for NeoVim which is
trivially easy to customize. This commit adds NvChad as a git submodule
at '~/.config/nvim/' and stores my custom configuration at
'~/.config/NvChad', so it must be manually symlinked on installation to
'~/.config/nvim/lua/custom' in order to work. The README.md is updated
to include this instruction.

My custom configuration is based on the default template that NvChad
prompts you to create initially, with very few additions, including bits
from my original '~/.vimrc', a few custom LSP servers, a few custom
keybindings, the "falcon" colorscheme, pseudo-transparency, and the
vim-tmux-navigator plugin, which makes Neovim and tmux panes more or
less interoperable, which is a great convenience.

[1]: https://nvchad.com