nvim: add clangd LSP and clang-format support
"clangd" and "clang-format" were added to the ensure_installed Mason
table. Additionally, the ".clang-format" file from the latest Linux
kernel tree was copied to "configs/" and is passed as an argument
whenever "clang-format" is invoked. Note that this suppresses the
reading of ".clang-format" from the current project's root, so this
setting must be overriden in a per-project basis as needed.