nvim: add settings for the Neovide GUI
Add a check to init.lua to determine whether we're running in the
Neovide[1] GUI for NeoVim. If so, read the Neovide settings file at
"configs/neovide.lua". Currently, this file only contains a "guifont"
setting, which is set to a Nerd Font compatible "SauceCodePro"; as well
as "wireframe" effect for the cursor, to visually display the change of
editing modes.

[1]: https://neovide.dev