neomutt: set options for listing of IMAP folders
The ‘imap_check_subscribed’ variable, according to the
(documentation)[1], “configures whether IMAP folder browsing will look
for only subscribed folders or all folders. This can be toggled in the
IMAP browser with the <toggle-subscribed> function.”

The effect is to list the relevant folders in the remote, which is
particularly useful for Gmail. The setting of this option in each
particular account's configuration file allows for additional
flexibility.

[1]: http://www.mutt.org/doc/manual/#imap-list-subscribed