neomutt: add “autoencrypt” script
As the name implies, this script — when loaded with something like
`neomutt -e 'source ~/.config/mutt/etc/autoencrypt.rc' …` — will cause
all sent emails to be signed and encrypted, if possible, by default. If
not, NeoMutt should™ prompt the user for a valid PGP key fingerprint,
and refuse to send the email unless provided with one manually.

This is most useful for use in scripts where E2EE is important — though
I can make no guarantees that it won't fail spectacularly and put you
out of a job. You have been warned!