nvim: set 'virtualedit' on blockwise visual mode
Per `:h 've`, virtual editing means that the cursor can be positioned
where there is no actual character. I've set this to "block", meaning
blockwise visual, or block selection (see `:h ^v`), meaning that, as
long as we're in V-BLOCK mode, we can move the cursor anywhere.

This is ideal for situations where you want to select, say, multiple
indented lines without the intentation. Normally, you'd only be able to
do so with blockwise visual if the final selected line happened to be
the longest. With 've=block', not only is this made easier, but it also
matches the blockwise selection behavior of the "tmux-yank" plugin for
tmux, which I already use very often.