README.md: remove leading '$' on shell code blocks
These make it hard to just copy and paste, though they do make it easier
to grep for these lines… but you definitely shouldn't be doing that
anyway. I mean… imagine grepping for shell commands in a markdown file!
Madness! In case this “feature” is indeed desired, it may be easier to
just use sed? Or maybe learn Perl! Up to you.