nvim: add another plugin for Go
This — the “gopher” plugin — adds some basic conveniences for Go. These
include adding json and yaml struct tags (mapped, respectively, to
`<leader>gsj` and `<leader>gsy`), wrapping the `go mod` and `go get`
commands with `:GoMod` and `:GoGet`, as well as automatically generating
tests with the `:GoTests…` family of commands, and many more. The plugin
is also configured to automatically install its own dependencies when
first installed or subsequently updated.