rTorrent: throttle upload and enable DHT and PEX
By default, the maximum upload rate is now 1024KB. This can, and should,
be increased as needed once rTorrent is launched and fully operational.
The rationale behind this change is in the fact that, in my
configuration, rTorrent starts very soon after boot, which, on my remote
machines, can cause the the entire bandwidth to be consumed immediately
by rTorrent. This default should cause seeding to not have any
noticeable impact in the network performance of other services.

In addition, DHT and Peer Exchange have been enabled by default and, for
DHT, a default node has been added. Finally, a default tracker has also
been added, and both it and the default DHT node are automatically added
to new torrents.