dwm: set tag labels to icons and layout to monocle
While it's easier to know how to “go” to a tag by knowing it's number,
it's way more useful to know *what* is in a tag. This change sets the
tag labels to icons from Nerd Fonts. These match my normal day-to-day
use of dwm, which is to have separate tags for, in order: terminals,
editors, browsers, video players, android emulation, games, music,
chatting, and finally password management.

On the change to the default layout, the biggest feature of dwm is its
convenience and ease of use, and the composability of work environments
through tiling windows is the biggest part of that. However, I very
often find myself separating all my activities into tags, and more often
than not using the monocle layout. This commit is written having that in
mind, though it's obviously still very much possible (and easy) to swith
to a tiling layout at will.