nvim: fix NvChad overriding jumplist navigation
From the Vim reference manual:

  A "jump" is a command that normally moves the cursor several lines
  away. If you make the cursor "jump" the position of the cursor before
  the jump is remembered.  You can return to that position with the "''"
  and "``" commands, unless the line containing that position was
  changed or deleted.  The following commands are "jump" commands: "'",
  "`", "G", "/", "?", "n", "N", "%", "(", ")", "[[", "]]", "{", "}",
  ":s", ":tag", "L", "M", "H" and the commands that start editing a new
  file.
  […]
  Jumps are remembered in a jump list.  With the CTRL-O and CTRL-I
  command you can go to cursor positions before older jumps, and back
  again.

See `:h jumplist` for more information.

As it happens, jumping around using only marks can be rather annoying,
and before installing NvChad (or even NeoVim), I was already used to
jumping around with ^o and ^i. Now, as you may know, a <C-i> is often
treated the same as a <Tab> character (ASCII value of 9). Some NeoVim
UIs however, like Neovide, and some terminals, like Kitty, can
distinguish between these two just fine, so this commit maps <C-i>
to perform jumplist navigation explicitly.

This makes it so that, in compatible UIs/terminals, <Tab> is used for
buffer navigation, and <C-i> is used for jumplist navigation. In
incompatible interfaces, a <C-i> will simply be treated as a <Tab>, and
thus used only for buffer navigation.