add ‘dmenusteam’
This script allows you to run Steam games under these directories:

	* /mnt/media/SteamLibrary/steamapps/
	* ~/.steam/steam/steamapps/

Non-games, such as Proton, the Steam runtime, etc., are filtered out,
and remaining titles can be selected with Dmenu. If Steam is already
running, the selected title will start and the script will exit. If
Steam is not running, it will start, and the selected title will run. In
the latter case, the script will watch Steam's child processes and, when
the Reaper exits (which should happen only if the game exits), Steam
will shut down, and the script will exit.

You should not use this script if you want multiple titles — such as a
game client and server — to run at the same time, unless you start Steam
before running the script.