HACK: rTorrent: disable sessions to avoid crashes
Currently, my rTorrent installation really likes to crash very
dramatically, dumping its core all over the place, messing up the
terminal, and, worst of all, annoying me when I'm already having a bad
day. If this keeps happening I might just have a gamer moment.

To avoid such a calamitous outcome, I commented out a line. The downside
is that session management ceases to work, so that it is impossible to
remember which torrents have been added and such.

TODO: send a strongly-worded letter to the rTorrent devs begging for
help.