Revert "tmux.conf: add vim-like bindings to switch panes"
This reverts commit dab22825668a710a07c3aefdfb60fe941e160ff3.

Apparently, one isn't supposed to manually set the pane-switching
bindings, because `christoomey/vim-tmux-navigator`. I had committed
dab2282 because the bindings weren't working. Turns out I'd forgotten to
actually install the plugins on the new system. 😅

Per the instructions,[^1] all I had to do is `<C-Space> + I` to install
them. (Maybe we ought to add this to the `README.md` so that I don't
make the same mistake whenever I install these settings to a new
system.)

[^1]: https://github.com/tmux-plugins/tpm/tree/99469c4#installing-plugins