rTorrent: re-enable sessions
We removed this on f6cbc80 because sessions caused crashes with
`rtorrent-ps-ch`, which I was using because regular `rtorrent-ps`
wouldn't compile on this system. As it turns out, *something has
happened* in the past 166 days that fixed rtorrent-ps. (No, I'm not
willing to learn how to use git-bisect(1) just to figure out exactly
when, ok?)