nvim: remove mini lib jump modules
`mini.jump` breaks the `,` mapping (see `:h ,`), which is at least as
useful as the `;` mapping, which is available! As far as I can tell,
there's no option to map something like a
`MiniJump.config.mappings.repeat_jump_backward`, per the docs. This
outweighs whatever benefit the plugin may add, by a lot.

As for `mini.jump2d`, I've just never found it to be that useful.
Perhaps I'll add vim-sneak if I need this kind of flexibility in the
future.