.zprofile: fix ssh-agent detection and initialization
This change refines the SSH agent setup by improving the detection
logic. The line `ssh-add -l > /dev/null 2>&1; echo $?` is used to check
the exit status of `ssh-add -l`. If the exit status is 2, it indicates
that the SSH agent is not running, prompting the script to start a new
agent and ensuring that the SSH agent is only initialized when
necessary. The graphical server startup command is also repositioned to
run conditionally, optimizing when it starts. Finally, `grep` away the
`echo` statement that's output by ssh-agent by default.