keepassmenu: add script to copy passwords from KeePassXC using dmenu
Similar to `passmenu2`, but reads from the KeePassXC file at
`~/Documents/Passwords.kdbx` using the `expect` scripts from 409fa0d.
Because a KeePassXC database may take a few seconds to be opened, we
cache the listing of entries for up to one hour, updating it earlier if
the database has been modified. The Perl script here is hacky, but
should™ properly parse the output of `keepass-cli`, provided some update
to KeePassXC doesn't modify the output format, which is vaguely
tree-like.

We also add a dwm keybinding to invoke this script.

In the future, I'd like to write something similar to wrap around
keepassxc-ssh-add with dmenu.