sunshine.service: unset DISPLAY and WAYLAND_DISPLAY env vars
This should force Sunshine to run in KMS mode, which:

  1. Prevents it from breaking in HyprLand, which can't identify the app
     that's requesting the screen capture permission, so we can't
     actually *see* anything when Moonlight connects (though, now that I
     think of it, maybe we could just have an `exec-once` line for
     sunshine in the HyprLand config).
  2. Allows us to use it on the display manager as well, so we can log
     in through it, even if we can't just magically `ioctl` into an
     agetty session (which is dumb anyways). Of course, this requires
     us to `loginctl enable-linger` ourselves, so that this unit
     actually runs when we're not logged in yet.

In fact, the only reason I even had autologin on getty@tty1 is because I
wanted my user units to run as soon as the computer booted. Now that
I've actually read through some of loginctl(1), I think that this was
always the wrong approach.